requirements.txt
numpy
pandas
Pillow
torch
torchvision
torchmetrics
kagglehub
matplotlib
loguru
